Add patch to address Bug#1066630
This bug was caused by a change in dpkg 1.22.6, that enabled
-Werror=implicit-function-declaration.  In dpkg version 1.22.6, the
compiler flag -Werror=implicit-function-declaration was enabled by
default for all architectures in build flags (in dpkg 1.22.5 it had
been enabled conditionally on the abi time64 feature, so only on most
32-bit architectures). The reason for doing this is avoiding
misbehavior of software affected by the ReleaseGoals/64bit-time
transition. When using a function with a time_t argument with an
implicit declaration, its type would be assumed to be int causing
silent truncation, or no symbol redirection (to their _time64
counterparts) would take effect.